Do not ignore PAGER env. var. if it does not contain pager args
The PAGER environmental variable was being ignored if it did not contain
arguments for the pager command. Typically, PAGER just contains the name
of the pager, e.g. "export PAGER=more", and pagerExecPath() was not
returning a valid path in that case.

Fix this by checking whether the environmental variable only contains
the name of the pager and returning an empty slice for the arguments.

Note that it looks like programs accept arguments after the name of the
pager, e.g. "export PAGER='more --lines 10'", and that is still being
handled (as it was before).